normalize GPIO configuration
rename GPIO's that have 'RPI' in the name to a more generic TARGET_GPIO
naming scheme. Additionally default spi_gpio_irq to a GPIO defined in
nrc-vendor.h so that it can easily be defaulted.

Note that the GPIO deifnes in nrc-vendor.h are appropriate for the
Rpi based NRC7292 EVB.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>